refactor: cy beforerun and afterun seed
